+++ .env.example
@@ -1,3 +1,8 @@
+# (required) Environment mode (local, staging, production)
+# Should be set to local when developers are running the API locally as it will disable some feature's side effects (e.g Freshdesk ticket creation)
+
+ENVIRONMENT_MODE=local
+
 # (required) Freshdesk
 # If the environment mode is set to local, you can left it empty as there will not be any Freshdesk request sent
