+++ src/config.ts
@@ -1,10 +1,23 @@
 import dotenv from "dotenv";
 
+export enum EnvironmentMode {
+  Local = "local",
+  Staging = "staging",
+  Production = "production",
+}
+
+// Load environment variables
+
 dotenv.config();
 
 // AWS SDK
 
 export const AWS_REGION: string = "ca-central-1";
+
+// Environment mode
+
+export const ENVIRONMENT_MODE: EnvironmentMode =
+  mapEnvironmentModeFromStringToEnum(loadRequiredEnvVar("ENVIRONMENT_MODE"));
 
 // Express
 
@@ -49,4 +62,18 @@ function loadRequiredEnvVar(envVarName: string): string {
   }
 
   return envVar;
+}
+
+function mapEnvironmentModeFromStringToEnum(
+  environmentMode: string,
+): EnvironmentMode {
+  if (environmentMode === "staging") {
+    return EnvironmentMode.Staging;
+  }
+
+  if (environmentMode === "production") {
+    return EnvironmentMode.Production;
+  }
+
+  return EnvironmentMode.Local;
 }
